requirements.txt
# Core dependencies
cryptography>=41.0.0
python-gnupg>=0.5.1
PyYAML>=6.0
requests>=2.31.0
click>=8.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
